global_routes: dont auto include hostname+env
Previously, I always included hostname+env. That may be the long-run
behavior we'd like, so that all logs have consistent metadata.

However, in the short-run for the migration, I'd rather just try to keep
data writing in exactly the way it did previously. This change allows
that to happen.

---

Moreover, this change also adds a new route (Gearcmd Duration), which
was recently added to log-router